Add Windows build scripts and accessibility fixes
Add Windows development helpers and multiple accessibility/compatibility fixes. Add build.bat and setup_dev_env.ps1 to simplify MSVC/Qt/MLT/vcpkg setup and build on Windows. CMake: add Qt Core5Compat, enable_testing(), adjust translations target name, and add vcpkg include workaround for pthread.h on Windows. Accessibility: use qrc: audio paths; enhance WelcomeWizard (strip HTML for accessible names, focus label, announce page body, delayed initial announce) and improve announcer usage in MainWindow export to avoid duplicate signal handlers. Prism integration: update screen reader code to new Prism v0.7.1 API (PrismConfig/PrismContext, backend initialization/shutdown, feature querying and name retrieval) and add m_context member. Misc fixes: change default project FPS to 25, add missing includes, and use explicit int cast for trackCount to avoid size_type mismatch. These changes improve Windows developer UX, update to a newer Prism API, and enhance accessibility behavior and runtime robustness.

+++ src/core/timeline.h
@@ -24,7 +24,7 @@ public:
 
     // Tracks
     [[nodiscard]] const QVector<Track *> &tracks() const { return m_tracks; }
-    [[nodiscard]] int trackCount() const { return m_tracks.size(); }
+    [[nodiscard]] int trackCount() const { return static_cast<int>(m_tracks.size()); }
     [[nodiscard]] Track *trackAt(int index) const;
 
     void addTrack(Track *track);
